Reorganize reloc_eq code a bit
